requirements.txt
python==3.7
Pytorch==1.4.0
Tensorflow==1.13.1
visdom==0.1.8.9
torchnet==0.0.4
pandas==1.1.4
tqdm==4.51.0
numpy==1.19.2
Pillow==8.0.1
pytz==2020.4
pyzmq==20.0.0